Implemented item upload.
Temp button that needs to be replaced with a material design type button in the Catalog Fragment.

Item class changed (added constructor and edited existing ones)

newItem ViewFragment is now calling the create from the item repo, item Repository has the uploadItem method which pushes a new slot into the database and adding a new key.